Move record comment impl into common directory
