+++ src/common/record/score.ts
